feat: codeup1030 추가

+++ course_student/java-project2/src/main/java/com/example/codeup/codeup1030.java
@@ -6,6 +6,6 @@ public class codeup1030 {
     public static void main(String[] args) {
         Scanner sc = new Scanner(System.in);
 
-        System.out.println(sc.nextInt());
+        System.out.println(sc.nextLong());
     }
 }
